log at global 01

+++ D_Solver.cuh
@@ -106,9 +106,9 @@ public:
         std::cout << "d" << std::endl;
         for (int i = 0; i < cloth->numConstraint; i++) {
             printf("d[%.3f, %.3f, %.3f]\n",
-                   d_d[3 * i + 0],
-                   d_d[3 * i + 1],
-                   d_d[3 * i + 2]);
+                   static_cast<float>(d_d[3 * i + 0]),
+                   static_cast<float>(d_d[3 * i + 1]),
+                   static_cast<float>(d_d[3 * i + 2]));
             printf("h[%.3f, %.3f, %.3f]\n",
                    h_d[3 * i + 0],
                    h_d[3 * i + 1],
